+++ server/src/interfaces/ICad.ts
@@ -13,6 +13,7 @@ interface ICad {
   live_map_url: string;
   steam_api_key: string;
   features: string[];
+  max_citizens: string;
 }
 
 export default ICad;